+++ src/UHC/Util/CHR/Solve/TreeTrie/MonoBacktrackPrio.hs
@@ -46,6 +46,8 @@ module UHC.Util.CHR.Solve.TreeTrie.MonoBacktrackPrio
   
   , chrSolve
   
+  , slvFreshSubst
+  
   , getSolveTrace
   
 {-
@@ -325,7 +327,7 @@ class ( IsCHRSolvable env cnstr guard bprio prio subst
       -- , MonadIO m -- for debugging
       , Fresh Int (ExtrValVarKey (VarLookupVal subst))
       -- , VarLookupKey subst ~ ExtrValVarKey cnstr
-      -- , ExtrValVarKey (VarLookupVal subst) ~ VarLookupKey subst
+      , ExtrValVarKey (VarLookupVal subst) ~ VarLookupKey subst
       , VarTerm (VarLookupVal subst)
       ) => MonoBacktrackPrio cnstr guard bprio prio subst env m
 
@@ -679,6 +681,7 @@ cvtSolverReductionStep (SolverReductionDBG pp) = return (SolverReductionDBG pp)
 -- | PP result
 ppSolverResult
   :: ( MonoBacktrackPrio c g bp p s e m
+     , VarUpdatable s s
      , PP s
      ) => Verbosity
        -> SolverResult s
@@ -698,7 +701,7 @@ ppSolverResult verbosity (SolverResult {slvresSubst = s, slvresResidualCnstr = r
                   >-< "Steps"   >-< indent 2 (vlist ss)
                | otherwise = Pretty.empty
     return $ 
-          "Subst"   >-< indent 2 s
+          "Subst"   >-< indent 2 (s `varUpd` s)
       >-< "Work"    >-< indent 2 (vlist ws)
       >-< pextra
 
@@ -1014,9 +1017,7 @@ chrSolve opts env = slv
           -- set prio for this alt
           sndl ^* chrbstBacktrackPrio =: bprio
           -- fresh vars for unbound body metavars
-          (freshSubst :: s) <- fmap (foldr (|+>) varlookupEmpty) $
-            forM (Set.toList $ varFreeSet body `Set.difference` freeHeadVars) $ \v ->
-              modifyAndGet (sndl ^* chrbstFreshVar) fresh >>= \v' -> return $ (varlookupSingleton v (varTermMkKey v') :: s)
+          freshSubst <- slvFreshSubst freeHeadVars body
           -- add each constraint from the body, applying the meta var subst
           newWkInxs <- forM body $ addConstraintAsWork . ((freshSubst |+> matchSubst) `varUpd`)
           -- mark this combi of chr and work as visited
@@ -1028,6 +1029,20 @@ chrSolve opts env = slv
           slv
 
     -- misc utils
+
+-- | Fresh variables in the form of a subst
+slvFreshSubst
+  :: forall c g bp p s e m x .
+     ( MonoBacktrackPrio c g bp p s e m
+     , ExtrValVarKey x ~ ExtrValVarKey (VarLookupVal s)
+     , VarExtractable x
+     ) => Set.Set (ExtrValVarKey x)
+       -> x
+       -> CHRMonoBacktrackPrioT c g bp p s e m s
+slvFreshSubst except x = 
+    fmap (foldr (|+>) varlookupEmpty) $
+      forM (Set.toList $ varFreeSet x `Set.difference` except) $ \v ->
+        modifyAndGet (sndl ^* chrbstFreshVar) (freshWith $ Just v) >>= \v' -> return $ (varlookupSingleton v (varTermMkKey v') :: s)
 
 -- | Lookup work in a store part of the global state
 slvLookup